+++ t/testdata/expected_output/plan_with_non_number.sh
@@ -1,0 +1,14 @@
+#!/bin/sh
+#
+# plan_with_non_number.sh
+# Make sure plan with a non-number number of tests fails
+# By J. Stuart McMurray
+# Created 20260317
+# Last Modified 20260317
+
+. ./t/t.subr
+. ./shmore.subr
+
+tap_plan "kittens"
+
+tap_done_testing